+++ src/GRA.Domain.Repository/IAvatarElementRepository.cs
@@ -7,7 +7,7 @@ namespace GRA.Domain.Repository
     public interface IAvatarElementRepository : IRepository<Model.AvatarElement>
     {
         Task<AvatarElement> GetByItemAndColorAsync(int item, int? color);
-        Task<ICollection<AvatarElement>> GetUserAvatarAsync(int userId);
+        Task<ICollection<AvatarElement>> GetUserAvatarAsync(int userId, int languageId);
         Task SetUserAvatarAsync(int userId, List<int> elementIds);
         void RemoveByItemId(int id);
     }